rename currentUser hook to useCurrentUser
