+++ components/pages/index/HomePage.test.js
@@ -5,7 +5,7 @@ import 'jest-styled-components';
 import renderWithApolloClient from '__tests__/helpers/renderWithApolloClient';
 import renderWithTheme from '__tests__/helpers/renderWithTheme';
 
-import { useCurrentUser } from 'lib/apollo/hooks/state/currentUser';
+import { useCurrentUser } from 'lib/apollo/hooks/state/useCurrentUser';
 
 import Homepage from 'pages/index';
