+++ lib/apollo/hooks/state/useCurrentUser.test.js
@@ -3,7 +3,7 @@ import { MockedProvider } from '@apollo/client/testing';
 
 import CurrentUser from 'graphql/queries/currentUser.graphql';
 
-import { useCurrentUser } from './currentUser';
+import { useCurrentUser } from './useCurrentUser';
 
 describe('useCurrentUser', () => {
   test('should return current user data', async () => {
